requirements.txt
gpiozero
pyyaml